+++ questions.csv
@@ -15,14 +15,14 @@ How many code completions and chat requests in GH Copilot Pro?, Unlimited, Unlim
 GH Copilot Pro cost?  $?, 10, $10 USD per month
 A specific type of code generation available in Pro but not Free?, test, Unit test generation is available in Copilot Pro but not in Copilot Free
 What do you get more of with GH Copilot Pro+?, Premium Requests, Premium requests and also Priority infrastructure access
-GH Copilot Business has centralized management for what controls?, management and policy, Centralized management for access & policy; additional security and compliance features.
+GH Copilot Business has what controls centralized?, management and policy, Centralized management for access & policy; additional security and compliance features.
 Another place you can use Chat listed for GH Copilot Business?, Mobile, Chat is obviously available for all versions but they say also Mobile for Business/Enterprise
 Two things they say GH CP Business has 'filtering' for ?/?, vulnerabilities/public code, Security Vulnerabilities and Public Code filtering; Code referencing for cases where public code is used or referenced.
 A kind of legal protection that you get with GH Copilot Business/Enterprise?, IP indemnity, Intellectual Property (IP) indemnity; Also enterprise-grade security; safety and privacy.
 What does Enterprise integrate with that Business does not? Github ? ?, Enterprise Cloud, Also includes personalized code suggestions based on internal/private code
 GHCP Enterprise supports AI-powered tags and summaries for?, Pull Requests, AI-powered tags and summaries for Pull Requests
 What does (only) GH Copilot Enterprise allow you to customize or fine-tune?, Models, Fine-tune or customize Copilot Models on your own codebase
-GHCP Enterprise can do this; and generate this; across your code base ?/?, search/documentation, AI-powered search and documentation generation across your codebase
+GHCP Enterprise can do this; and generate this; across your code base ?/?, search/documentation, GH Copilot Enterprise has: AI-powered search and documentation generation across your codebase
 GHCP Enterprise provides these throughout the Github platform?, Action Buttons, it also provides integration into GitHub as a chat interface; so developers can converse about their codebase.
 What key and the right arrow will accept part of an inline suggestion?, Control, Control + Right Arrow; it goes about a word at a time
 What key to accept the full inline suggestion?, Tab, Tab key; The course says the right arrow key will do it but it does not work for me
